Readme.md
Multiple Question Paper Generator is a file based C++ program that loads questions from a file and then creates random question papers. The input is taken the count of question papers and the number of questions per paper. Increase the number of questions or add more questions to avoid repeating the same question in multiple question papers.